Replace *HTTPS Everywhere* with *HTTPS by default*.
The former only applies by default to a closed list of sites, unless the user
adds a (not trivial) rule to its config, or it enables EASE.  In the
meanwhile, when providing just a host name or a plain HTTP URL in CENO's
location bar, Ouinet's Origin mechanism may successfully retrieve a *block
page*, which is quite confusing for a tool like CENO.

The later extension, though not recent nor sophisticated, does the job for
most cases relevant to CENO.

An alternative would be to force plain HTTP to go through the Injector
mechanism, but this has important drawbacks.  For instance, if example.com is
a site with working HTTPS and HTTP->HTTPS redirection (as probably the vast
majority of sites out there, or at least most of interest to CENO):

1. User enters `example.com` at the location bar (or follows
   `http://example.com/` link).

2. Extension/Ouinet see that it is plain HTTP and send the request to
   Injector.  Injector fetches and signs response.

3. Ouinet starts announcing `example.com` (which covers
   `{http,https}://(www.)example.com/`).

4. The response is an HTTP->HTTPS redirection, CENO follows it.

5. `https://example.com` happens not to be blocked, so it fetches it via
   Origin, i.e. it does not cache nor announce its content.

So now we would have a client announcing `example.com` that others may reach
when looking up the dcache, only to find that it only seeds the redirection
and no content from the front page.  In other words, it would add noise to the
dcache for a lookup that could have been resolved instantly locally by doing
the automatic HTTP->HTTPS upgrade.

Also, if the later client wanted to access `https://example.com/` directly, it
would still find and ask the former client for a URL that it is not seeding.

So without deeper/wider changes to how redirections and DHT groups are
handled, sending HTTP always to Injector would be detrimental for the reduced
set of sites that it would serve.